Created the MasterGlyph class to hold data of a high-resolution version of the font. Made the TextRendererAa struct contain an array of MasterGlyphs. Removed GlyphAa and moved most of its code into MasterGlyph. Made the GDI render and bitmap storage in MasterGlyph both 1-bit deep. Previously 32-bit bitmap data was used. This is to save RAM. The grayscale data for the glyph pixels doesn't matter much in MasterGlyph because grey-scale anti-aliased versions of the font will be generated by downsizing the 1-bit data from MasterGlyph.
Current status is that the 1-bit GDI render works and the MasterGlyph creates a correct and un-GDI'd version of the data. There is no code to create the scaled-down glyphs. The kerning calculations (which will now be done using the high-res MasterGlyph) might work, but the code is currently commented out.