Add DfFontSource as a mechanism to allow in-memory fonts to be queried for font sizes etc, similar to how ListFontSizesInFile() works for on-disk fonts. (These two mechanisms should be unified in the future).
Make in memory fonts not be declared static in the .h files because that causes them to be duplicated by every compilation unit that #includes them. Instead declare them extern and add .cpp file to contain non-static definitions.

Declared the in-memory fonts as const, because they are.

Modify LoadFontFromMemory() to take a const * parameter.